move MetadataInfo
